test: split basic DBF memo lifecycle regressions
