+++ tests/CMakeLists.txt
@@ -4523,6 +4523,7 @@ copperfin_apply_windows_stack_reserve(test_studio_host)
 
 copperfin_add_test(test_dbf_table cf_vfp_assets
     test_dbf_table.cpp
+    test_dbf_table_memo_lifecycle.cpp
 )
 
 copperfin_add_test(test_dbf_text_encoding cf_vfp_assets
